pyproject.toml
[project]
name = "chexpert-dinov3"
version = "0.1.0"
description = "CheXpert 2023 multi-label classification via fine-tuned DINOv3 ViT-H+/16"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.7,<2.12",
    "torchvision>=0.22",
    "transformers>=4.56",
    "pillow>=10",
    "pandas>=2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "tqdm>=4.66",
    "pyyaml>=6",
]

# The cluster's NVIDIA driver is 12.8 (12080); torch built against CUDA
# 13.0 (the default on PyPI) refuses to see the GPU. Pin torch/torchvision
# to wheels from the cu128 index.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
